fix: comment clean ups
