+++ src/common/decorators/controller/index.ts
@@ -1,2 +1,1 @@
 export * from './Controller';
-export * from './AllowAnonymous';
