+++ src/common/decorators/route/RouteMetadata.ts
@@ -1,7 +1,7 @@
 export interface RouteMetadata {
   path: string;
-  method: 'get' | 'post' | 'put' | 'delete' | 'patch';
+  method: 'POST';
   methodName: string;
 }
 
-export const ROUTE_METADATA_KEY = Symbol('route');
+export const ROUTE_METADATA_KEY = 'routes';
