+++ src/common/middleware/index.ts
@@ -1,3 +1,2 @@
 export * from './ValidationPipe';
 export * from './ValidationPipeException';
-export * from './AuthenticationMiddleware';
